review migratio and checked in checkbox

+++ database/migrations/2021_11_10_103921_create_reviews_table.php
@@ -16,7 +16,7 @@ class CreateReviewsTable extends Migration
         Schema::create('reviews', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->tinyInteger('vote');
-            $table->unsignedInteger('restaurant_id');
+            $table->unsignedBigInteger('restaurant_id');
             $table->foreign('restaurant_id')->references('id')->on('restaurants');
             $table->timestamps();
         });
